Slowdown fixed, profiling added, high-res DISABLED. See more details.
The slowdown only occurs in Chrome to my knowledge, where it did not prior to 7/16-7/18 or so 2013. Something in a Chrome update must have resulted in much more limited canvas caching capabilities (maybe some speed optimization for which our many-canvases is an edge-case). Regardless, I've disabled high-res caching for now, and though the problem was also occurring at low resolutions, I implemented a fix for that by ignoring the caching of sections without data.

As a part of debugging the issue I added a simple timer for profiling.
Usage:
time.start("a timer");
//do stuff
time.stop("a timer");
time.report();